requirements.txt
azure-core==1.33.0
azure-functions==1.23.0
certifi==2025.4.26
charset-normalizer==3.4.1
faiss-cpu==1.11.0
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
txtai==8.5.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
